Update 2.2.0: Runtime editing
Added the runtime editor! It is a new app on the diagnostic tablet where you can swipe the door with the admin (red) card, and change certain settings (except passes atm) of a single door or all doors of a multidoor. After you click the screen, the door automatically updates with the server and updates itself with the new settings. In case you made a mistake, it does save a copy of the old settings in dbsettings.txt (i believe that's it)

All commit names: 

* Work on new editing ability of diagnostic tablet

* Magreader lights can be controlled (new OS update)

* Finished prototype for runtime editing

* UUID resseting button design upgrade

* Runtime editing bug fixes #1

Moved decrypting to only be for forceopen, because thats all that does.

* Diagnostic bug fixes: term lines/magreader lights

Magreader numbers upped by 1 (change to code)
Door control, if single door, wouldn't be changed until reboot (fixed)

* diagnostic major snag

Gotta figure out how to get the diagnostic tablet to read it right. I think its an issue with it getting references, not the actual data of values.

* Added deepcopy and possible runtime edit bugfix

* diagnostics bug fixes

Forgot a .. as well as changing pageChange to editChange. Gotta figure out why multidoors receive multiple changesettings messages

* final diagnostic breaking bugfixes

Fixed the bugs in the diagnostic mode that caused it to not work. Almost ready for release

* entire door bug fixes on diagnostic

Entire door diagnostic didn't work, so I feexed it hopefully

* Update README.md