RC final bug fixes #2 and server logging
Doors forgot the reader address parameter when receiving a remote control signal. The remote control update should be ready for release. More features are being added before release however.
Added in another command to the server to allow external devices to send in logs for the server to display on the screen (so remote control events can be received & listed)