securityAPI program seperation & server checks
SecurityAPI setup function now tells the server it is a custom type of door (instead of a single) This will prevent the door from showing up in the remote control menu since it isn't supported for rc.
Server checks through its list of available commands before performing it's functions (it would crash if any other command was given) I also added in a way to add more commands via function in runtime to allow for modules.