Server modules, sector,s, and more #1
Added "support" for making modules for the server to perform your own actions. This includes a prefab for your own programs in the extras/server module prefab/ directory.
Started work on a sector module that works with the server (although, I hard coded some of the sector code into the server itself)
Made server check if modules folder is there and download the sector module automatically the first time (since it might be a required module)
Added sectorcontrol program with no code in it to the extras folder, so I can work on it later. This will be what you connect redstone to to lock down sectors and such.

The only reason I am making all the sector code in a seperate module is to try and find everything that is necessary for a module to work. I really REALLY hope this works :)