Database sector management prototype
All styles & Localizations are done for the new sector system as well as how it works. This is a very experimental build and is ready for testing. The chances of this working first time are close to 0. I worked so hard on this and hope it doesn't go to waste :smiley: