Sector module checks doors
In order for a door to know a user's sector status, it has to receive it from the server. The door now will send another message before the checkPasses message to check if the door is allowed to open. Some reworking will be done to what is sent back to the door, however, as it doesn't tell the door what the current sector status is as well. One issue I'll have to figure out as well is how to reduce time it takes to open doors (as the current sector system causes it to wait twice for modem messages, causing it to take twice as long.)

Version 2.4.1 might just be me moving all the sector code BACK into the main code, or at least integrating it into the  checkPasses program, since it sends the same data as checkPasses anyway. God, it's so confusing :/