Sector refactoring & reader light changes
Sector work has been "completed" in the server I believe. The server sends the message to the sector module now instead of the door having to send it. This should mean the delay between swiping a card and receiving a response will be low for all the doors instead of sector doors taking forever.
Fixed MineOS database bugs (certain variables were incorrect)
Added functionality to allow the door to disable a lockdown if a user swipes twice in 3 seconds (if valid pass as well.)
Added functonality where if you have the pass that lets you bypass lockdown doors, you still must have valid passes for that specific door too.
Created a prototype function that changes the lights on magreaders. Instead of creating a thread every time a door requests it, it will have one constantly running thread that reads through an array of magreaders. This should mean no overlapping commands and potentially less lag if lots of doors are swiped, but i don't know if it works yet.

I swear this update is going to take the LONGEST to debug, as I have not been testing any of this. I am only going to test once I believe everything has been finished. Me brain can'ts handle this!