Began work on sectorcontrol
Move over key functions to sectorcontrol program, which will be able to control the sectors. It will also include the ability to set certain sector executables to bundled redstone lines. This would allow for the ability to automatically control certain lockdowns with redstone lines. I plan on "maybe" adding ability to control from the gui itself, but to make life easier on myself and less complicated for myself as well, the gui will only be for setting redstone lines to bundled redstone colors :)