Sector extreme prototype finished
This, is a buggy prototype that likely will not work
Dear God...
There's more
No...

Through a few hours of constant pain, sweat, tears, and blood (maybe not blood though), I managed to "finish" the system. i can't really think of anything else that must be added before testing, so it's time for testing. yaaaaay...
Another thing I made was rangeextender, because I plan on adding the ability to use a linking card in doorcontrol systems, so your server can send messages to doorcontrols outside of the range of modems and in other dimensions, if thats what you like...