Experimental Server GUI changes
Updated the server's writing to be a little more refined. History scrolls up neatly in it's own area. Depending on the size of the screen and tier, the viewport is different, so the amount of history displayed is more or less. History can also scroll up as lines are added, meaning info can be displayed above and below history area without being removed.