doorTable online check
Due to the nature of remoteControl using the doorTable to open doors, if a door is offline, removed, modem is changed, or swapped to linking card, the doorlist doesn't change, causing extra doors to be there and not work.
Every time the server is started, it sends a message to every single door control asking if it is online. If no response is heard, it removes the door from the list and saves it.