doorcontrol hotfix & readMe
Multidoor redstone doors didnt work with remote control. It's fixed now.
Updated magreader lights to take in a new var called check that, after the door is finished iterating through a lighting table, will set the light's color to. This "Should" make it so if a door is on sector lockdown, the red light will come back.
After a door restart, it should check if door sectors are fixed (as in, if a sector is locked down, locked open, etc.