Server "finish", sector and pass modules done, etc
Server has been finished I believe. Still have to change around the code for the other programs before testing anything cause of the sheer amount of changes.
Sector module was rewritten to allow for the new way to handle sectors with modules. This involves rewrites to other programs as well that are needed. Also fixed it so it sends updated list to sectorcontrol now when changed.
Passes module was "successfully' moved from the old server. I think it is all there, but will need testing.
Changed module prefab to add new modulemsg part.