Diagnostic & Sectorcontrol update
Changed both diagnostic and sectorcontrol programs to work with new setup of server.
Diagnostics had to change it's query to ask for both passSettings and sectors, then change how they read the query because of the change in table formatting.
Sectorcontrol had to add a new table for sectorStatus and send the sectorStatus only instead of all the sectors. Also changed query to only grab the sectors and sectorStatus.