"Finished" Passes module for MineOS
I moved all the code from the old database to the module if it was relevant and changed around some of the code. However, I do still have to change the userTable stuff to check the right parts of it. After that, a test may be in order. I pray that everything works first try :thumbsup: