Bug Fixes #3
Fixed server not sending the table because of an issue with query.
Fixed server not saving changed table after refactoring of old table.
Put a return in passes module for server to see if it fixed something (it did not)
Fixed userlist growing in size after removing it to add new users.
Fixed database not saving the query's .data variable as the userTable (thats where it's sent)
Fixed database's updateServer function not initializing data as a table.

Database "seems" to be working fine, with users and such. However, I can't freaking test it due to the server only allowing one message through, then disregarding any more. I have no idea what is wrong with the server and will work on it next.