Remote Control and MOOOOORE
* Created Remote Control module

Remote Control Module Beta ready

* Reset button to door setup module to clear doors

* Comments on quikidlink

For info and to allow for others to kinda understand :)

* diagnostic sectorcontrol and ipairs

1. Replaced all occurrences of pairs with ipairs since its stupid not to. We absolutely want to maintain order now, do we?
2. Implemented and converted sectorcontrolgui program that was pulled in from another user to the diagnostictablet as part of it. This'll allow for easier testing of sectors and such.
3. Added forceprogram variable to config file to let you force a program to startup. This is nice for situations where you only want one program to run (eg: sectorcontrolgui on a desktop instead of tablet)

* Delete sectorcontrolgui.lua

Not needed

* Lots of work on RuntimeDoorEditing

Made progress on readding reader modification to the system. Almost finished the work on checking the component types of the devices (needs keypad support). To make it work it has to resetup all doorcontrol and reader blocks in the system for advanced (rfid) systems.

* Finished experimental door editing again

* IPairs bugfix (revert)

Turns out IPairs is STUPID >:(

reverted all changes to ipairs as it didn't workie

* Remote control bugfixes

Pretty much finished remote control bug fixes. Seems to work fineish

* Door editing finished & minor fixes

1. Finished door editing and made it worked
2. Fixed Sector Control on the diagnostic tablet
3. yes :)

* Prepared to release

Preparing to release changes!